+++ src/aqua_diagnostics/timeseries/plot_timeseries.py
@@ -70,7 +70,7 @@ class PlotTimeseries(PlotBaseMixin):
         # Filling them
         self.get_data_info()
 
-    def run(self, var: str, units: str = None, region: str = None, outputdir: str = './',
+    def run(self, var: str, units: str = None, outputdir: str = './',
             rebuild: bool = True, dpi: int = 300, format: str = 'png'):
         """
         Run the PlotTimeseries class.
@@ -78,7 +78,6 @@ class PlotTimeseries(PlotBaseMixin):
         Args:
             var (str): Variable name to be used in the title and description.
             units (str): Units of the variable to be used in the title.
-            region (str): Region to be used in the title and description.
             outputdir (str): Output directory to save the plot.
             rebuild (bool): If True, rebuild the plot even if it already exists.
             dpi (int): Dots per inch for the plot.
@@ -88,11 +87,10 @@ class PlotTimeseries(PlotBaseMixin):
         self.logger.info('Running PlotTimeseries')
         data_label = self.set_data_labels()
         ref_label = self.set_ref_label()
-        description = self.set_description(region=region)
-        title = self.set_title(region=region, var=var, units=units)
+        description = self.set_description()
+        title = self.set_title(var=var, units=units)
         fig, _ = self.plot_timeseries(data_labels=data_label, ref_label=ref_label, title=title)
-        region_short = region.replace(' ', '').lower() if region is not None else None
-        self.save_plot(fig, var=var, description=description, region=region_short, rebuild=rebuild,
+        self.save_plot(fig, var=var, description=description, region=self.region, rebuild=rebuild,
                        outputdir=outputdir, dpi=dpi, format=format)
         self.logger.info('PlotTimeseries completed successfully')
 
@@ -105,6 +103,7 @@ class PlotTimeseries(PlotBaseMixin):
         - AQUA_catalog
         - AQUA_model
         - AQUA_exp
+        - AQUA_region
         - std_startdate
         - std_enddate
         """
@@ -114,10 +113,13 @@ class PlotTimeseries(PlotBaseMixin):
                 self.catalogs = [d.AQUA_catalog for d in data]
                 self.models = [d.AQUA_model for d in data]
                 self.exps = [d.AQUA_exp for d in data]
+                # We expect all data arrays to have the same region
+                self.region = data[0].AQUA_region if hasattr(data[0], 'AQUA_region') else None
                 break
         self.logger.debug(f'Catalogs: {self.catalogs}')
         self.logger.debug(f'Models: {self.models}')
         self.logger.debug(f'Experiments: {self.exps}')
+        self.logger.debug(f'Region: {self.region}')
 
         # TODO: support ref list
         for ref in [self.ref_monthly_data, self.ref_annual_data]:
@@ -125,45 +127,41 @@ class PlotTimeseries(PlotBaseMixin):
                 self.ref_catalogs = ref.AQUA_catalog
                 self.ref_models = ref.AQUA_model
                 self.ref_exps = ref.AQUA_exp
+                self.logger.debug(f'Reference: {self.ref_catalogs} {self.ref_models} {self.ref_exps}')
                 break
-        self.logger.debug(f'Reference: {self.ref_catalogs} {self.ref_models} {self.ref_exps}')
 
         for std in [self.std_monthly_data, self.std_annual_data]:
             if std is not None:
                 self.std_startdate = std.std_startdate if std.std_startdate is not None else None
                 self.std_enddate = std.std_enddate if std.std_enddate is not None else None
+                self.logger.debug(f'Standard deviation dates: {self.std_startdate} - {self.std_enddate}')
                 break
-        self.logger.debug(f'Standard deviation dates: {self.std_startdate} - {self.std_enddate}')
 
-    def set_title(self, region: str = None, var: str = None, units: str = None):
+    def set_title(self, var: str = None, units: str = None):
         """
         Set the title for the plot.
 
         Args:
-            region (str): Region to be used in the title.
             var (str): Variable name to be used in the title.
             units (str): Units of the variable to be used in the title.
 
         Returns:
             title (str): Title for the plot.
         """
-        title = super().set_title(region=region, var=var, units=units, diagnostic='Time series')
+        title = super().set_title(region=self.region, var=var, units=units, diagnostic='Time series')
         return title
 
-    def set_description(self, region: str = None):
+    def set_description(self):
         """
         Set the caption for the plot.
         The caption is extracted from the data arrays attributes and the
         reference data arrays attributes.
         The caption is stored as 'Description' in the metadata dictionary.
 
-        Args:
-            region (str): Region to be used in the caption.
-
         Returns:
             description (str): Caption for the plot.
         """
-        description = super().set_description(region=region, diagnostic='Time series')
+        description = super().set_description(region=self.region, diagnostic='Time series')
         return description
 
     def plot_timeseries(self, data_labels=None, ref_label=None, title=None):